TRIPLEX-846 ChipSelectDropdown | Исправить выравнивание Dropdown относительно таргета
