+++ stories/NumberField/examples/DefaultExample.tsx
@@ -4,10 +4,12 @@ import { NumberField, EComponentSize, EFormFieldStatus } from "@sberbusiness/tri
 export const DefaultExample = () => {
     const [value, setValue] = useState<string>("");
 
-    const handleInputChange: React.ChangeEventHandler<HTMLInputElement> = (event) => setValue(event.target.value);
+    const handleInputChange: React.ChangeEventHandler<HTMLInputElement> = (event) => {
+        setValue(event.target.value);
+    };
 
     return (
-        <div style={{ maxWidth: "300px" }}>
+        <div style={{ maxWidth: 300 }}>
             <NumberField
                 size={EComponentSize.LG}
                 status={EFormFieldStatus.DEFAULT}
