+++ stories/NumberField/examples/index.ts
@@ -1,3 +1,5 @@
+export * from "./PlaygroundExample";
+export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
 export * from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export * from "./SizesExample";
