+++ stories/release-notes/v1/1.24.0.mdx
@@ -16,6 +16,10 @@ import { Meta, Title, Heading } from "@storybook/addon-docs/blocks";
 
 - Исправлен адаптивный вид `CheckboxYGroup`.
 
+**ChipSelect, ChipSort**
+
+- Исправлено позиционирование `Dropdown`: теперь выравнивается по левому краю таргета.
+
 **Col**
 
 - Уменьшен горизонтальный отступ между колонками: `@grid-horizontal-gap` изменён с `12px` на `8px`.
